refactor: remove a lot of base58 and ECDSA stuff from bitcoin.py
