+++ lib/transaction.py
@@ -23,6 +23,7 @@
 import bitcoin
 from bitcoin import *
 from util import print_error
+from util_coin import var_int
 import time
 import struct
 import chainparams
