+++ setup.py
@@ -96,12 +96,14 @@ setup(
     data_files=data_files,
     py_modules=[
         'chainkey.account',
+        'chainkey.base58',
         'chainkey.bitcoin',
         'chainkey.blockchain',
         'chainkey.bmp',
         'chainkey.chainparams',
         'chainkey.commands',
         'chainkey.daemon',
+        'chainkey.eckey',
         'chainkey.i18n',
         'chainkey.interface',
         'chainkey.mnemonic',
@@ -117,6 +119,7 @@ setup(
         'chainkey.synchronizer',
         'chainkey.transaction',
         'chainkey.util',
+        'chainkey.util_coin',
         'chainkey.verifier',
         'chainkey.version',
         'chainkey.wallet',
